layer1_CNN/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0
tqdm>=4.66.0
matplotlib>=3.8.0
opencv-python>=4.9.0
onnx>=1.16.0
